subscriptions work for strings, ints, and arrays of ints of known length (defined in yaml)... now to not break those things while handling arrays of unknownlength (first case of stuff other than strings with length prepended)

+++ ROS#/Messages/geometry_msgs/Point32.cs
@@ -15,8 +15,8 @@ namespace Messages.geometry_msgs
 
 		public class Point32
 		{
-			public double x;
-			public double y;
-			public double z;
+			public float x;
+			public float y;
+			public float z;
 		}
 }
