+++ ROS#/Messages/std_msgs/Float32.cs
@@ -15,6 +15,6 @@ namespace Messages.std_msgs
 
 		public class Float32
 		{
-			public double data;
+			public float data;
 		}
 }
